train/segmentation/baseline/pyproject.toml
[project]
name = "baseline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "segmentation-models-pytorch>=0.5.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "timm>=0.9.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "pytorch-lightning>=2.5.5",
    "fastai>=2.8.5",
]
